+++ src/include/d2tmc.h
@@ -30,6 +30,7 @@
 #include "gameobjects/cSpecialInfos.h"
 #include "gameobjects/cUpgradeInfo.h"
 #include "gameobjects/units/cUnitInfos.h"
+#include "gameobjects/structures/cStructureInfo.h"
 
 #include <memory>
 #include <array>
